8.gün

+++ Entities/Concrete/Car.cs
@@ -10,8 +10,8 @@ namespace Entities.Concrete
         public int Id { get; set; }
         public int BrandId { get; set; }
         public int ColorId { get; set; }
-        public decimal ModelYear { get; set; }
-        public int DailyPrice { get; set; }
+        public int ModelYear { get; set; }
+        public decimal DailyPrice { get; set; }
         public string Description { get; set; }
 
     }
